Do not run doctests on Python 2
Making these doctests all work on both Python 2 and 3 is not really
useful, and could result in examples that are confusing to the reader.
For example, some contain functions that are invoked differently in
Python 2 and 3, and including monkeypatch code to rectify this is bad
for documentation purposes.

Python 3 is our main priority.