Allow subs that are missing a double blank newline
This often happens on subs which were first created as a mixed language
subtitle, for example chs/eng, and then were stripped using naive methods (such
as ed/sed) that don't understand newline preservation rules in SRT files.

This means that when you are, say, only keeping chs, and the line only contains
english, you end up with not only no content, but also all of the content lines
are stripped instead of retaining a newline.